+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `LERN` will be documented in this file.
 
+### 2.0.0
+- Added a `LERN` facade
+- Monolog\Logger Support
+    - With the ability to use custom Logger and Handler instances
+- Custom table name for migration (see the new config file)
+
 ### 1.1.0
 - Added Slack